requirements.txt
Django
djangorestframework
mediapipe
numpy
opencv_python
ultralytics
ffmpeg-python
ffmpeg
django-cors-headers
